pyproject.toml
[tool.poetry]
name = "pedidos-env"
version = "0.1.0"
description = ""
authors = ["Dados Abertos de Feira <dadosabertosdefeira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
python-decouple = "^3.4"
django-public-admin = "^0.0.5"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
